Changes transformations
Made it so the image size doesn't change just the smilies. So a skinny
smily has the same image dimesnsions
Made it so that images have a padding of white pixels as per james request
because otherwise the data is usuable.
Reduced the skews so that our data is more consistent